feature: Tidying up Svelte implementation
- fixes the weird bug with recipe loading
